# Add sumCache