@caio 2018-11-10
Looks good, thanks! 

Things are fresh in my head so everything makes sense, but soon I think it'll be a bit more difficult. Do you mind adding some docs on what is its purpose and how it works? Elaborating on what I think will confuse me in the future:

* `sumCache.valid` reads like a boolean to me - does `lastValidIndex` or something similar make more sense?
* Only the checks for < 4 and the bit shifts make it more obvious that this is only about every 4th index; I can't think of a clearer name than `sumCache` so maybe just a comment on top of the struct definition suffices
* The `nil` checks on every method felt weird to me (first time seeing a struct being used like this) but it's quite smart, I like it! Just maybe a comment on top of the struct definition is missing, like saying it only gets created after having more than 100 distinct samples in the digest