@vmihailenco 2018-11-11
I forgot to `Invalidate` cache in `summary.setAt` which skewed benchmarks. With fixed `summary.setAt` caching does not look worth it any more. Sorry for the noise :)